I made the following changes to visit.f
1) Aleks recommended to change my write functions to use the format function, so I change them.

2) If VisIt failed to connect to the Nek5000 sim, the sim would stop running. 
This has been fix and the sim will continue in the current run state, which is to run by default.

3) Also a problem with processing VisIt commands would change the runflag which could result 
in the Nek5000 sim stopping. This has been fixed.

4) If VisIt disconnects from the Nek5000 sim for any reason, the sim would stop running. 
I have change the code so the sim will continue running.

5) The Nek5000 sim would stop at the end of the sim if compiled with VisIt even if VisIt was 
not connected. I have change the code to the following. If VisIt is not connect to the Nek5000 
sim when the visit_end is called the sim will exit. If VisIt is connect it will stop and 
wait for the VisIt client to exit the sim.




git-svn-id: https://svn.mcs.anl.gov/repos/nek5/trunk/nek@953 24248562-e74d-0410-ab4b-d4d80c744c73