Fixed dsavg_acc
The problem was that ACC DATA PRESENT(mg_imask) failed at runtime, with
an error saying that the array wasn't present on device, even though it
is copied-in in acc_copy_all_in().  We noticed that mg_imask is set as
an "equivalence" to mg_mask, and added mg_mask to the COPYIN statement.
This appeared to fix the problem.