Began implementing gop_acc (#303)
* Began implementing gop_acc()

The single-node allreduce is a null op, so we've eliminated the
host/device data copy.  The multi-node allreduce will need to either (a)
do and update host/device and run MPI on CPU or (b) use GPU direct
without host/device update.  For now, I have implemented (a).

* Began implementing glsc3_acc

* Applied glsc3_acc to calls in hmh_gmres

A call to glsc3_acc inside an ACC kernel had to be inlined, even when
glsc3_acc was compiled as an ACC ROUTINE.  I received errors about
unsupported nested compute w/ that call.

* Fixed compiler error w/ gfortran