only initialize bx in nekuic if ifmhd=.true.
This avoids initializing bx in non-MHD cases where ldimt>npscal+3. We do this to allow diagnostic use of the T array and to make it easier to visualize diagnostic quantities (like the artificial viscosity fields in CMT-nek). This also avoids breaching the bounds of BX if ldimt>npscal+3 and lbelt=1